Separate catalog and catalog details actions
