+++ src/services/actions/catalog-detail.js
@@ -1,0 +1,3 @@
+// Actions of catalog detail
+export const ACTION_CATALOG_DETAIL_SHOW = 'ACTION_CATALOG_DETAIL_SHOW';
+export const ACTION_CATALOG_DETAIL_HIDE = 'ACTION_CATALOG_DETAIL_HIDE';
