+++ src/services/actions/catalog.js
@@ -1,8 +1,6 @@
 import { getIngredientsService } from '../../api/ingredients';
 
 // Actions of catalog
-export const ACTION_CATALOG_DETAIL_SHOW = 'ACTION_CATALOG_DETAIL_SHOW';
-export const ACTION_CATALOG_DETAIL_HIDE = 'ACTION_CATALOG_DETAIL_HIDE';
 export const ACTION_CATALOG_REQUEST = 'ACTION_CATALOG_REQUEST';
 export const ACTION_CATALOG_REQUEST_FAIL = 'ACTION_CATALOG_REQUEST_FAIL';
 export const ACTION_CATALOG_REQUEST_SUCCESS = 'ACTION_CATALOG_REQUEST_SUCCESS';
