+++ src/services/reducers/index.js
@@ -2,9 +2,11 @@ import { combineReducers } from 'redux';
 import { constructorReducer } from './constructor';
 import { orderReducer } from './order';
 import { catalogReducer } from './catalog';
+import { catalogDetailReducer } from './catalog-detail';
 
 export const rootReducer = combineReducers({
   cart: constructorReducer,
   order: orderReducer,
   catalog: catalogReducer,
+  catalogDetail: catalogDetailReducer,
 });
